[MINOR] Fix edge case tests to reflect new changes

+++ src/test/java/org/apache/sysds/test/component/frame/FrameCustomTest.java
@@ -35,7 +35,7 @@ public class FrameCustomTest {
 		double maxp1 = Integer.MAX_VALUE + 1.0;
 		MatrixBlock mb = TestUtils.generateTestMatrixBlock(100, 100, maxp1, maxp1, 1.0, 23);
 		FrameBlock f = DataConverter.convertToFrameBlock(mb);
-		assertTrue(f.getSchema()[0] == ValueType.INT64);
+		assertTrue(f.getSchema()[0] == ValueType.FP64);
 	}
 
 	@Test
@@ -50,8 +50,7 @@ public class FrameCustomTest {
 	public void castErrorValue() {
 		MatrixBlock mb = new MatrixBlock(10, 10, Double.parseDouble("2.572306572E9"));
 		FrameBlock f = DataConverter.convertToFrameBlock(mb);
-		assertTrue(f.getSchema()[0] == ValueType.INT64);
-
+		assertTrue(f.getSchema()[0] == ValueType.FP64);
 	}
 
 	@Test
